requirements.txt
discord.py==2.1.0
async-timeout==4.0.2
python-dotenv~=0.21.0